+++ AAEmu.Game/Models/Game/Transfers/TransferBindingDoodads.cs
@@ -5,7 +5,7 @@
         public uint Id { get; set; }
         public uint OwnerId { get; set; }
         public string OwnerType { get; set; }
-        public int AttachPointId { get; set; }
+        public sbyte AttachPointId { get; set; }
         public uint DoodadId { get; set; }
     }
 }
